Cast issue

+++ NewAPI/Controllers/ParkingLotController.cs
@@ -8,8 +8,8 @@ public class ParkinglotRequest
     public string Address { get; set; }
     public int Capacity { get; set; }
     public int Reserved { get; set; }
-    public decimal Tariff { get; set; }
-    public decimal DayTariff { get; set; }
+    public double Tariff { get; set; }
+    public double DayTariff { get; set; }
 }
 
 public class LicenseplateRequest
